refactor: no persisting session
Unfortunately, persisting session would require to get all cookies but 
we can't get the secure cookies as it'd be a XSS attack. Instead, we'll 
have to login everytime we run the scraping, which is not that bad.

Another change is that now the browser gets initialzied before asking 
for credentials, so it feels faster.