feat: configurable wait until
To load a page, we need `domcontentloaded` but for images, we need to 
wait for network to be done.

If we only use `domcontentload` then some pictures won't load correctly.

And if we use 'networkidle0' then when the dom is too big it won't load 
on time after network gets idle.